move RouterError
