+++ crates/ibc/src/mock/ics18_relayer/error.rs
@@ -1,6 +1,5 @@
-use crate::core::ics03_connection;
 use crate::core::ics24_host::identifier::ClientId;
-use crate::core::ics26_routing::error::RouterError;
+use crate::core::{ics03_connection, RouterError};
 use crate::Height;
 use displaydoc::Display;
